se agrega el countdown en presentación
