+++ frontend/src/reunion/Reunion.js
@@ -84,6 +84,22 @@ class Reunion extends React.Component {
     return null;
   }
 
+  segundosRestantes = () => {
+    const tema = this.temaSeleccionado();
+    if (tema.inicio === null) {
+      return tema.cantidadDeMinutosDelTema * 60;
+    }
+    let tiempo = Date.now();
+    if (tema.fin !== null) tiempo = Date.parse(tema.fin);
+    return Math.round(tema.cantidadDeMinutosDelTema * 60
+        - (tiempo - Date.parse(tema.inicio)) / 1000);
+  }
+
+  temaActivo = () => {
+    const { inicio, fin } = this.temaSeleccionado();
+    return inicio !== null && fin === null;
+  }
+
   render() {
     const VistaSeleccionada = this.obtenerVista();
     // TO DO: Ver qué se debería mostrar en caso de carga o error
@@ -94,7 +110,9 @@ class Reunion extends React.Component {
         <ReunionContainer>
           <VistaSeleccionada tema={this.temaSeleccionado()}
             terminarTema={this.terminarTema}
-            empezarTema={this.empezarTema} />
+            empezarTema={this.empezarTema}
+            segundosRestantes={this.segundosRestantes()}
+            temaActivo= {this.temaActivo()}/>
           <Sidebar handleSelection={this.handleSelection}
             selectedElement={this.state.selectedElement}
             link={this.temaSeleccionado().linkDePresentacion} />
